+++ src/ansibleguy-webui/aw/views/job.py
@@ -7,7 +7,7 @@ from aw.utils.http import ui_endpoint_wrapper, ui_endpoint_wrapper_kwargs
 from aw.model.job import JobExecution, JobExecutionResultHost
 from aw.api_endpoints.job_util import get_viewable_jobs
 from aw.utils.util import get_next_cron_execution_str
-from aw.views.forms.job import job_edit, job_repository_edit, job_credentials_edit
+from aw.views.forms.job import job_edit, job_repository_edit, job_credentials_edit, job_clone
 
 LIMIT_JOB_RESULTS = 10
 LIMIT_JOB_LOG_RESULTS = 50
@@ -71,6 +71,7 @@ urlpatterns_jobs = [
     path('ui/jobs/credentials/<int:credentials_id>', job_credentials_edit),
     path('ui/jobs/credentials', job_credentials),
     path('ui/jobs/log', job_logs),
+    path('ui/jobs/manage/job/clone/<int:job_id>', job_clone),
     path('ui/jobs/manage/job/<int:job_id>', job_edit),
     path('ui/jobs/manage/job', job_edit),
     path('ui/jobs/repository/<int:repo_id>', job_repository_edit),