+++ src/ansibleguy-webui/aw/templatetags/util.py
@@ -3,7 +3,7 @@ from functools import cache
 from django import template
 
 from aw.config.main import config
-from aw.settings import STATIC_URL
+from aw.settings import STATIC_URL, AUTH_MODE
 from aw.config.navigation import NAVIGATION
 from aw.utils.version import get_version as get_version_util
 
@@ -27,6 +27,11 @@ def get_logo() -> str:
 @register.simple_tag
 def set_var(val):
     return val
+
+
+@register.filter
+def auth_sso(_) -> bool:
+    return AUTH_MODE == 'saml'
 
 
 @register.filter
@@ -118,3 +123,48 @@ def find(data: str, search: str) -> bool:
         data = str(data)
 
     return data.find(search) != -1
+
+
+# see: https://github.com/grafana/django-saml2-auth/blob/main/django_saml2_auth/errors.py
+SAML_ERROR_CODES = {
+    1100: 'EMPTY_FUNCTION_PATH',
+    1101: 'PATH_ERROR',
+    1102: 'IMPORT_ERROR',
+    1103: 'GENERAL_EXCEPTION',
+    1104: 'CREATE_USER_ERROR',
+    1105: 'GROUP_JOIN_ERROR',
+    1106: 'NO_REVERSE_MATCH',
+    1107: 'ERROR_CREATING_SAML_CONFIG_OR_CLIENT',
+    1108: 'NO_SAML_RESPONSE_FROM_CLIENT',
+    1109: 'NO_SAML_RESPONSE_FROM_IDP',
+    1110: 'NO_NAME_ID_IN_SAML_RESPONSE',
+    1111: 'NO_ISSUER_IN_SAML_RESPONSE',
+    1112: 'NO_USER_IDENTITY_IN_SAML_RESPONSE',
+    1113: 'NO_TOKEN_SPECIFIED',
+    1114: 'NO_USERNAME_OR_EMAIL_SPECIFIED',
+    1115: 'SHOULD_NOT_CREATE_USER',
+    1116: 'INACTIVE_USER',
+    1117: 'NO_METADATA_URL_OR_FILE',
+    1118: 'NO_SAML_CLIENT',
+    1119: 'NO_JWT_ALGORITHM',
+    1120: 'INVALID_METADATA_URL',
+    1121: 'NO_METADATA_URL_ASSOCIATED',
+    1122: 'INVALID_REQUEST_METHOD',
+    1123: 'CANNOT_DECODE_JWT_TOKEN',
+    1124: 'USER_MISMATCH',
+    1125: 'NO_JWT_SECRET',
+    1126: 'NO_JWT_PRIVATE_KEY',
+    1127: 'NO_JWT_PUBLIC_KEY',
+    1128: 'INVALID_JWT_ALGORITHM',
+    1129: 'NO_USER_ID',
+    1130: 'INVALID_TOKEN',
+    1131: 'INVALID_NEXT_URL',
+}
+
+
+@register.filter
+def saml_error_by_code(error_code: int) -> str:
+    if error_code not in SAML_ERROR_CODES:
+        return ''
+
+    return f" ({SAML_ERROR_CODES[error_code]})"