+++ src/ansibleguy-webui/aw/urls.py
@@ -4,11 +4,13 @@ from django.conf.urls import include
 from django.contrib import admin
 from django.contrib.auth import views as auth_views
 
+
 from web_serve_static import urlpatterns_static
 from aw.api import urlpatterns_api
 from aw.views.main import urlpatterns_ui, catchall, logout
 from aw.config.environment import check_aw_env_var_true
 from aw.utils.deployment import deployment_dev
+from aw.config.environment import auth_mode_saml
 
 urlpatterns = []
 
@@ -16,6 +18,20 @@ if deployment_dev() or check_aw_env_var_true(var='serve_static', fallback=True):
     urlpatterns += urlpatterns_static
 
 urlpatterns += urlpatterns_api
+
+if auth_mode_saml():
+    from django_saml2_auth import views as saml_auth_views
+    urlpatterns += [
+        re_path(r'^a/saml/', include('django_saml2_auth.urls')),
+        re_path(r'^a/sso/$', saml_auth_views.signin),
+        re_path(r'^o/$', saml_auth_views.signout),
+    ]
+
+else:
+    urlpatterns += [
+        re_path(r'^o/$', logout),
+    ]
+
 urlpatterns += [
     # auth
     path('a/', include('django.contrib.auth.urls')),  # login page
@@ -23,6 +39,7 @@ urlpatterns += [
     path('_admin/', admin.site.urls),  # admin page
     path('o/', logout),
 ]
+
 urlpatterns += urlpatterns_ui
 urlpatterns += [
     # fallback