+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   * Auto-Completion via Tab-Key
   * Select using Up/Down/Enter Keys
 * Ability to clone existing jobs
+* Ability to sort jobs on manage UI
 
 ----