+++ CHANGELOG.md
@@ -8,7 +8,7 @@
   * Auto-Completion via Tab-Key
   * Select using Up/Down/Enter Keys
 * Ability to clone existing jobs
-* Ability to sort jobs on manage UI
+* Ability to sort jobs and repositories
 * Split-up Repository Forms
 
 ----